+++ src/view/utils/index.js
@@ -1,4 +1,4 @@
-/* globals newspack_popups_view */
+/* globals gtag, newspackPopupsData, newspack_popups_view */
 
 /**
  * WordPress dependencies
@@ -184,3 +184,53 @@ export const parseOnHandlers = onAttributeValue =>
 		.split( ';' )
 		.filter( Boolean )
 		.map( onHandler => /(?<action>\w*):(?<id>(\w|-)*)\.(?<method>.*)/.exec( onHandler ).groups );
+
+/**
+ * Get all prompts on the page.
+ *
+ * @return {Array} Array of prompt elements.
+ */
+export const getPrompts = () => {
+	return [ ...document.querySelectorAll( '.newspack-popup' ) ];
+};
+
+/**
+ * Get raw prompt ID number from element ID name.
+ *
+ * @param {string} id Element ID of the prompt.
+ *
+ * @return {number} Raw ID number from the element ID.
+ */
+export const getRawId = id => {
+	const parts = id.split( '_' );
+	return parseInt( parts[ parts.length - 1 ] );
+};
+
+/**
+ * Get a GA4 event payload for a given prompt.
+ *
+ * @param {string} action      Action name for the event.
+ * @param {number} promptId    ID of the prompt
+ * @param {Object} extraParams Additional key/value pairs to add as params to the event payload.
+ *
+ * @return {Object} Event payload.
+ */
+export const getEventPayload = ( action, promptId, extraParams = {} ) => {
+	if ( ! newspackPopupsData || ! newspackPopupsData[ promptId ] ) {
+		return false;
+	}
+
+	return { ...newspackPopupsData[ promptId ], ...extraParams, action };
+};
+
+/**
+ * Send an event to GA4.
+ *
+ * @param {Object} payload   Event payload.
+ * @param {string} eventName Name of the event. Defaults to `np_prompt_interaction` but can be overriden if necessary.
+ */
+export const sendEvent = ( payload, eventName = 'np_prompt_interaction' ) => {
+	if ( 'function' === typeof gtag && payload ) {
+		gtag( 'event', eventName, payload );
+	}
+};